refactor(tests): reorganize suite into domain folders
Move checker, DNS resolver, and output test modules into dedicated tests/checker, tests/dns, and tests/output packages, relocate related support modules, and update imports while preserving full test behavior and coverage.
