+++ tests/checker/test_checker_core.py
