+++ tests/checker/test_checker_logging.py
