+++ tests/checker/test_checker_spf_policy.py
@@ -4,7 +4,7 @@ from provider_check.checker import DNSChecker
 from provider_check.dns_resolver import DnsLookupError
 from provider_check.status import Status
 
-from tests.checker_spf_support import make_provider_with_spf, make_spf_config
+from tests.checker.spf_support import make_provider_with_spf, make_spf_config
 from tests.support import BASE_PROVIDER, FakeResolver
