+++ tests/checker/test_checker_srv_optional.py
@@ -4,7 +4,7 @@ from provider_check.checker import DNSChecker
 from provider_check.provider_config import SRVRecord
 from provider_check.status import Status
 
-from tests.checker_srv_support import make_provider_with_srv, make_srv_config
+from tests.checker.srv_support import make_provider_with_srv, make_srv_config
 from tests.support import FakeResolver
 
 OPTIONAL_AUTODISCOVER = {
