+++ tests/checker/test_checker_srv_required.py
@@ -2,7 +2,7 @@ from provider_check.checker import DNSChecker
 from provider_check.provider_config import SRVRecord
 from provider_check.status import Status
 
-from tests.checker_srv_support import make_provider_with_srv, make_srv_config
+from tests.checker.srv_support import make_provider_with_srv, make_srv_config
 from tests.support import FakeResolver
 
 PRIMARY_REQUIRED = {
