+++ tests/checker/test_checker_txt_errors.py
@@ -2,7 +2,7 @@ from provider_check.checker import DNSChecker
 from provider_check.dns_resolver import DnsLookupError
 from provider_check.status import Status
 
-from tests.checker_txt_support import make_provider_with_txt, make_txt_config
+from tests.checker.txt_support import make_provider_with_txt, make_txt_config
 from tests.support import FakeResolver
