+++ tests/checker/test_checker_txt_required.py
@@ -1,7 +1,7 @@
 from provider_check.checker import DNSChecker
 from provider_check.status import Status
 
-from tests.checker_txt_support import make_provider_with_txt, make_txt_config
+from tests.checker.txt_support import make_provider_with_txt, make_txt_config
 from tests.support import FakeResolver
