+++ tests/checker/txt_support.py
