+++ tests/output/test_output_color.py
